+++ CHANGELOG.md
@@ -16,12 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - if present, coBib will store the journal in its elongated form
     - a user can then automatically convert to abbreviated forms during exporting (see the new `--abbreviate` and `--dotless` arguments)
 - Tentative Windows support by disabling the TUI
+- a `--preserve-files` argument was added to the following commands. Unless it is given, these will delete/rename associated files of affected entries:
+    - `DeleteCommand`
+    - `EditCommand`
+    - `ModifyCommand`
 
 ### Changed
 - the `modifications` of a `ModifyCommand` get interpreted as f-strings (#77,!37):
     - available variables are the entry's label and data fields
 - any `ListCommand` filter gets interpreted as a regex pattern (#76)
-- unless `--preserve-files` is passed to `DeleteCommand`, associated file will also be deleted
 
 ### Deprecated
 - the `ID` filter argument was fully replaced by `label`, unifying the CLI API. Support will be dropped in v3.3.0